Unity 6000.0.0b16. Fixed camera materials and missing noise texture when no camera source.

+++ Assets/SaveLoader.cs
@@ -400,7 +400,7 @@ namespace SS.Resources {
         surveillanceSourceEntities[i] = entity;
 
         entityManager.AddComponentData(entity, new SurveillanceSource() {
-          CameraIndex = i
+          CameraIndex = (byte)i
         });
       }
